refactor(Promotion): Promotion 도메인으로 수정
행사 목록을 정의해 두기 위해 enum 클래스로 생각하였지만 파일 입출력을 통해 행사 목록을 불러온다는 요구 사항에 맞게 수정하였습니다.
